core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain-anthropic>=1.3.0",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-xai>=1.1.0",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
